refactor layout filetree into pagesManifest endpoint
