+++ sites/example-project/src/pages/+layout.svelte
@@ -1,66 +1,4 @@
 <!-- This get's shipped with the template -- don't do local imports from $lib -->
-<script context="module">
-	// Import pages and create an object structure corresponding to the file structure
-	const pages = import.meta.glob(['/src/pages/*/**/+page.md']);
-	let pagePaths = Object.keys(pages).map((path) => path.replace('/src/pages/', ''));
-
-	// Create a tree structure from the array of paths
-	let fileTree = {
-		label: 'Home',
-		href: '/',
-		children: {},
-		isTemplated: false
-	};
-	pagePaths.forEach(function (path) {
-		path.split('/').reduce(function (r, e) {
-			if (e === '+page.md') {
-				let href = path.includes('[') ? undefined : encodeURI('/' + path.replace('/+page.md', ''));
-				return (r['href'] = href);
-			} else {
-				let label = e.includes('[') ? undefined : e.replace(/_/g, ' ').replace(/-/g, ' ');
-				r.isTemplated = e.includes('[');
-				return (
-					r?.children[e] ||
-					(r.children[e] = {
-						label,
-						children: {},
-						href: undefined,
-						isTemplated: false
-					})
-				);
-			}
-		}, fileTree);
-	});
-
-	// Recursively delete nodes and children nodes that don't have a label
-	function deleteEmptyNodes(node) {
-		if (node.children) {
-			Object.keys(node.children).forEach(function (key) {
-				deleteEmptyNodes(node.children[key]);
-				if (!node.children[key].label && !node.children[key].href) {
-					delete node.children[key];
-				}
-			});
-		}
-	}
-
-	deleteEmptyNodes(fileTree);
-
-	// Convert children objects into arrays of objects
-	function convertChildrenToArray(node) {
-		if (node.children) {
-			node.children = Object.keys(node.children).map(function (key) {
-				return node.children[key];
-			});
-			node.children.forEach(function (child) {
-				convertChildrenToArray(child);
-			});
-		}
-	}
-
-	convertChildrenToArray(fileTree);
-</script>
-
 <script>
 	import '@evidence-dev/tailwind/fonts.css';
 	import '../app.css';
@@ -82,6 +20,10 @@
 	$: if ($navigating) {
 		mobileSidebarOpen = false;
 	}
+
+	export let data;
+
+	let fileTree = data?.pagesManifest;
 </script>
 
 <!-- QueryStatus doesn't actually create any UI, just supplies toasts to ToastWrapper -->
